+++ src/base/text_format_enum.hh
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, Timothy Stack
+ * Copyright (c) 2025, Timothy Stack
  *
  * All rights reserved.
  *
@@ -27,30 +27,36 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "string_attr_type.hh"
+#ifndef lnav_text_format_enum_hh
+#define lnav_text_format_enum_hh
 
-#include "config.h"
+#include <sys/types.h>
 
-constexpr string_attr_type<void> SA_ORIGINAL_LINE("original_line");
-constexpr string_attr_type<void> SA_BODY("body");
-constexpr string_attr_type<ui_icon_t> SA_HIDDEN("hidden");
-constexpr string_attr_type<void> SA_REPLACED("replaced");
-constexpr string_attr_type<intern_string_t> SA_FORMAT("format");
-constexpr string_attr_type<void> SA_REMOVED("removed");
-constexpr string_attr_type<void> SA_PREFORMATTED("preformatted");
-constexpr string_attr_type<std::string> SA_INVALID("invalid");
-constexpr string_attr_type<std::string> SA_ERROR("error");
-constexpr string_attr_type<int64_t> SA_LEVEL("level");
-constexpr string_attr_type<int64_t> SA_ORIGIN_OFFSET("origin-offset");
-constexpr string_attr_type<text_format_t> SA_QUOTED_TEXT("quoted-text");
+#include "base/enum_util.hh"
 
-constexpr string_attr_type<role_t> VC_ROLE("role");
-constexpr string_attr_type<role_t> VC_ROLE_FG("role-fg");
-constexpr string_attr_type<text_attrs> VC_STYLE("style");
-constexpr string_attr_type<const char *> VC_GRAPHIC("graphic");
-constexpr string_attr_type<block_elem_t> VC_BLOCK_ELEM("block-elem");
-constexpr string_attr_type<styling::color_unit> VC_FOREGROUND("foreground");
-constexpr string_attr_type<styling::color_unit> VC_BACKGROUND("background");
-constexpr string_attr_type<std::string> VC_HYPERLINK("hyperlink");
-constexpr string_attr_type<ui_icon_t> VC_ICON("icon");
-constexpr string_attr_type<ui_command> VC_COMMAND("command");
+enum class text_format_t : uint8_t {
+    TF_BINARY,
+    TF_C_LIKE,
+    TF_JAVA,
+    TF_JSON,
+    TF_LOG,
+    TF_MAKEFILE,
+    TF_MAN,
+    TF_MARKDOWN,
+    TF_PYTHON,
+    TF_PCRE,
+    TF_RUST,
+    TF_SQL,
+    TF_XML,
+    TF_YAML,
+    TF_TOML,
+    TF_DIFF,
+    TF_SHELL_SCRIPT,
+    TF_LNAV_SCRIPT,
+    TF_UNKNOWN,
+};
+
+constexpr auto text_format_count
+    = lnav::enums::to_underlying(text_format_t::TF_UNKNOWN) + 1;
+
+#endif