+++ src/styling.hh
@@ -102,6 +102,9 @@ struct lnav_theme {
     positioned_property<icon_config> lt_icon_log_level_critical;
     positioned_property<icon_config> lt_icon_log_level_fatal;
 
+    positioned_property<icon_config> lt_icon_play;
+    positioned_property<icon_config> lt_icon_edit;
+
     positioned_property<style_config> lt_style_identifier;
     positioned_property<style_config> lt_style_text;
     positioned_property<style_config> lt_style_alt_text;